nv50/exa: reemit surface state in hostdata UTS if a flush occurs
This is just so the bufmgr knows we used the surface again, unlike
earlier chips nv5x doesn't require us to resubmit the source point
etc. again :)

How this looks is closer to how all the flush hooks should look.  We
have per-app contexts, it's a shame to do full re-emits.  Though,
the impact seems rather minimal so far.